Get information about available types of data processes (DataProcessTypeID)

Get information about available types of data processes (DataProcessTypeID)

get_dataprocesstype(save_file = FALSE, ...)

Arguments

save_file

Logical. Choose whether or not to save the data in an external .Rdata file in the working directory. Default: FALSE.

...

Other arguments that might define the path to data. All arguments accept a numeric code which is interpreted as the code of the specific product requested. Alternatively, you can supply the equivalent product name as a string which is case insensitive (see examples). Handle with care, this is important! For a list of all options available, see the parameters for each endpoint at https://popdiv.dfs.un.org/Demodata/swagger/ui/index#/

Examples

if (FALSE) {
D <- get_dataprocesstype()
D[, c("PK_DataProcessTypeID", "Name", "ShortName")]
}