Get information about available indicators (IndicatorID) and indicatortypeids (IndicatorTypeId)

Get information about available indicators (IndicatorID) and indicatortypeids (IndicatorTypeId)

get_iitypes(save_file = FALSE, ...)

Arguments

save_file
Logical. Choose whether or not to save the data in an external .Rdata file in the working directory. Default: FALSE;

...
Other arguments that might define the path to data. All arguments accept a numeric code which is interpreted as the code of the specific product requested. Alternatively, you can supply the equivalent product name as a string which is case insensitive (see examples). Handle with care, this is important! The following options are available:

dataProcessIds -- Data process ID as defined by the UNPD. Run the get_dataprocess function to see the available options;

startYear -- Start year. Default: NULL;

endYear -- End year. Default: NULL;

indicatorTypeIds -- Indicator type ID as defined by the UNPD. Run the get_indicatortypes function to see the available options;

isComplete -- isComplete is set to `2` ('Total') by default on the API;

locIds -- Location ID as defined by the UNPD. Run the get_locations function to see the available options;

locAreaTypeIds -- Location area type ID as defined by the UNPD. Run the get_locationtypes function to see the available options;

subGroupIds -- SubGroup ID as defined by the UNPD. Run the get_subgroups function to see the available options;

addDefault -- Logical. Default: FALSE;

includeDependencies -- Logical. Default: FALSE;

includeFormerCountries -- Logical. Default: FALSE.

Examples

if (FALSE) {
I <- get_iitypes()
I

only_itypes <- get_iitypes(indicatorTypeIds = 6)
only_itypes

only_iids <- get_iitypes(indicatorIds = 229)
only_iids

components <- get_iitypes(componentIds = 4)
components
}