Download data from UNPD portal

Download data from UNPD portal

get_recorddata(save = FALSE, verbose = TRUE, ...)

Arguments

save
Logical. Choose whether or not to save the data in an external .Rdata file in the working directory. Default: FALSE;

verbose
Whether to print the translated query from strings to digits for faster queries. By default set to TRUE.

...
Other arguments that might define the path to data. All arguments accept a numeric code which is interpreted as the code of the specific product requested. Alternatively, you can supply the equivalent product name as a string which is case insensitive (see examples). Handle with care, this is important! The following options are available:

dataProcessIds -- Data process ID as defined by the UNPD. Run the get_dataprocess function to see the available options;

startYear -- Start year. Default: NULL;

endYear -- End year. Default: NULL;

indicatorTypeIds -- Indicator type ID as defined by the UNPD. Run the get_indicators function to see the available options;

isComplete -- isComplete;

locIds -- Location ID as defined by the UNPD. Run the get_locations function to see the available options;

locAreaTypeIds -- Location area type ID as defined by the UNPD. Run the get_locationtypes function to see the available options;

subGroupIds -- SubGroup ID as defined by the UNPD. Run the get_subgroups function to see the available options;

addDefault -- Logical. Default: FALSE;

includeDependencies -- Logical. Default: FALSE;

includeFormerCountries -- Logical. Default: FALSE.

Examples

if (FALSE) {
#  You can provide all strings, all codes, or a combination of both
Y <- get_recorddata(dataProcessIds = "Census",
                   indicatorTypeIds = 8, # and support numeric of string names
                   locIds = "egypt", # all arguments are case insensitive
                   locAreaTypeIds = "Whole area",
                   subGroupIds = "Total or All groups",
                   isComplete = "Abridged")

head(Y)

# Same thing only with codes
X <- get_recorddata(dataProcessIds = 2,   # Census
                   indicatorTypeIds = 8, # Population by age and sex - abridged 
                   locIds = 818,         # Egypt
                   locAreaTypeIds = 2,   # Whole area 
                   subGroupIds = 2,      # Total or All groups
                   isComplete = 0)       # Age Distribution: Abridged

head(X)
}