Get information about available details for a given series of data

Get information about available details for a given series of data

get_seriesdata(save = FALSE, ...)

Arguments

save
Logical. Choose whether or not to save the data in an external .Rdata file in the working directory. Default: FALSE;

...
Other arguments that might define the path to data. All arguments accept a numeric code which is interpreted as the code of the specific product requested. Alternatively, you can supply the equivalent product name as a string which is case insensitive (see examples). Handle with care, this is important! The following options are available:

dataProcessIds -- Data process ID as defined by the UNPD. Run the get_dataprocess function to see the available options;

startYear -- Start year. Default: NULL;

endYear -- End year. Default: NULL;

indicatorTypeIds -- Indicator type ID as defined by the UNPD. Run the get_indicators function to see the available options;

isComplete -- isComplete;

locIds -- Location ID as defined by the UNPD. Run the get_locations function to see the available options;

locAreaTypeIds -- Location area type ID as defined by the UNPD. Run the get_locationtypes function to see the available options;

subGroupIds -- SubGroup ID as defined by the UNPD. Run the get_subgroups function to see the available options;

addDefault -- Logical. Default: FALSE;

includeDependencies -- Logical. Default: FALSE;

includeFormerCountries -- Logical. Default: FALSE.

Examples

# NOT RUN {
# You can provide all strings, all codes, or a combination of both
G <- get_seriesdata(dataProcessIds = 0:15, # possible processes
                   indicatorTypeIds = 25,    # M[x]
                   locIds = "Egypt",             # Egypt
                   locAreaTypeIds = c("whole area", "rural", "urban"), # all possible types
                   subGroupIds = 2)
G
# }