Wrapper for Life Table Methods

Wrapper for Life Table Methods

do_lifetable(X, verbose = TRUE, ...)

Arguments

X

Input data. UN format.

verbose

Logical value. If TRUE messages are printed as the method is applied. Set verbose = FALSE to silent the function.

...

Other arguments to be passed on to other methods and functions.

Value

A data.frame having the same number of columns as input data. Different numbers of rows. UN format.

See also

Examples

mx <- DDSQLtools.data$Mx5
do_lifetable(X = mx)
#> Additional (optional) arguments to control the output for the DemoTools::lt_abridged method:
#> Deaths Exposures nMx nqx lx AgeInt radix axmethod a0rule Sex IMR region mod SRB OAnew extrapLaw extrapFrom extrapFit ...
#> Check ?lt_abridged for details and default values.
#> # A tibble: 180 × 13
#>    LocID ReferencePeriod DataValue AgeStart AgeSpan SexID IndicatorID DataTypeID
#>    <int>           <int>     <dbl>    <int>   <int> <dbl> <chr>       <chr>     
#>  1   328            1993   0.0297         0       1     1 nMx         do_lifeta…
#>  2   328            1993   0.00665        1       4     1 nMx         do_lifeta…
#>  3   328            1993   0.00343        5       5     1 nMx         do_lifeta…
#>  4   328            1993   0.00378       10       5     1 nMx         do_lifeta…
#>  5   328            1993   0.00707       15       5     1 nMx         do_lifeta…
#>  6   328            1993   0.0134        20       5     1 nMx         do_lifeta…
#>  7   328            1993   0.0184        25       5     1 nMx         do_lifeta…
#>  8   328            1993   0.0223        30       5     1 nMx         do_lifeta…
#>  9   328            1993   0.0257        35       5     1 nMx         do_lifeta…
#> 10   328            1993   0.0361        40       5     1 nMx         do_lifeta…
#> # … with 170 more rows, and 5 more variables: DataTypeName <chr>,
#> #   AgeLabel <dbl>, AgeMid <dbl>, AgeEnd <dbl>, AgeID <dbl>