Download data from UNPD portal

Download data from UNPD portal

get_recorddata(save_file = FALSE, verbose = TRUE, ...)

Arguments

save_file
Logical. Choose whether or not to save the data in an external .Rdata file in the working directory. Default: FALSE;

verbose
Whether to print the translated query from strings to digits for faster queries. By default set to TRUE.

...
Other arguments that might define the path to data. All arguments accept a numeric code which is interpreted as the code of the specific product requested. Alternatively, you can supply the equivalent product name as a string which is case insensitive (see examples). Handle with care, this is important! The following options are available:

dataProcessTypeIds -- Data process Type ID as defined by the UNPD. Run the get_dataprocess function to see the available options;

dataProcessIds -- Data process ID as defined by the UNPD. Run the get_dataprocessid function to see the available options;

startYear -- Start year. Default: NULL;

endYear -- End year. Default: NULL;

indicatorTypeIds -- Indicator type ID as defined by the UNPD. Run the get_indicatortypes function to see the available options;

indicatorIds -- Indicator ID as defined by the UNPD. Run the get_iitypes function to see the available options;

isComplete -- isComplete is set to `2` ('Total') by default on the API;

locIds -- Location ID as defined by the UNPD. Run the get_locations function to see the available options;

locAreaTypeIds -- Location area type ID as defined by the UNPD. Run the get_locationtypes function to see the available options;

subGroupIds -- SubGroup ID as defined by the UNPD. Run the get_subgroups function to see the available options;

addDefault -- Logical. Default: FALSE;

includeDependencies -- Logical. Default: FALSE;

includeFormerCountries -- Logical. Default: FALSE.

Details

get_recorddata allows the user to supply string names for all arguments that have equivalent get_* functions. For example, get_iitypes for indicatorIds. The string used for all of these arguments should be the one from the column Name in the response from the get_* functions.

Once the data is read from the API, some transformations are applied:

Columns AreaName, DataReliabilityName, SubGroupName, DataStatusName, DataTypeName, DataTypeGroupName, IndicatorName, LocName, LocAreaTypeName, LocTypeName, ModelPatternName, ModelPatternFamilyName, PeriodGroupName, PeriodTypeName, RegName, SexName, StatisticalConceptName, SubGroupTypeName are converted to labelled factors with labelled

TimeStart and TimeEnd are returned with format 'DD/MM/YYYY'

Examples

if (FALSE) {
#  You can provide all strings, all codes, or a combination of both
Y <- get_recorddata(dataProcessTypeIds = "Census",
                   indicatorTypeIds = 8, # and support numeric of string names
                   locIds = "egypt", # all arguments are case insensitive
                   locAreaTypeIds = "Whole area",
                   subGroupIds = "Total or All groups",
                   isComplete = "Abridged")

head(Y)

# Same thing only with codes
X <- get_recorddata(dataProcessTypeIds = 2,   # Census
                   indicatorTypeIds = 8, # Population by age and sex - abridged 
                   locIds = 818,         # Egypt
                   locAreaTypeIds = 2,   # Whole area 
                   subGroupIds = 2,      # Total or All groups
                   isComplete = 0)       # Age Distribution: Abridged

head(X)
}