API Link Generator Function

API Link Generator Function

linkGenerator(
  server = getOption("unpd_server", "http://24.239.36.16:9654/un3/api/"),
  type,
  verbose_print,
  ...
)

Arguments

server
The path to the database. Check if the "unpd_server" option is is set. If not, defaults to "http://24.239.36.16:9654/un2/api/"

type
Type of data. Various options are available.

...
Other arguments that might define the path to data. All arguments accept a numeric code which is interpreted as the code of the specific product requested. Alternatively, you can supply the equivalent product name as a string which is case insensitive (see examples). Handle with care, this is important! The following options are available:

dataProcessTypeIds -- Data process Type ID as defined by the UNPD. Run the get_dataprocess function to see the available options;

dataProcessIds -- Data process ID as defined by the UNPD. Run the get_dataprocessid function to see the available options;

startYear -- Start year. Default: NULL;

endYear -- End year. Default: NULL;

indicatorTypeIds -- Indicator type ID as defined by the UNPD. Run the get_indicatortypes function to see the available options;

indicatorIds -- Indicator ID as defined by the UNPD. Run the get_iitypes function to see the available options;

isComplete -- isComplete is set to `2` ('Total') by default on the API;

locIds -- Location ID as defined by the UNPD. Run the get_locations function to see the available options;

locAreaTypeIds -- Location area type ID as defined by the UNPD. Run the get_locationtypes function to see the available options;

subGroupIds -- SubGroup ID as defined by the UNPD. Run the get_subgroups function to see the available options;

addDefault -- Logical. Default: FALSE;

includeDependencies -- Logical. Default: FALSE;

includeFormerCountries -- Logical. Default: FALSE.

Details

The link generator is based on the structure of the database created by Dennis Butler (in late 2018). To change the server used to make the requests, set this at the beginning of your script: options(unpd_server = "fill this out").

When requesting data from the structured data format (usually called from get_recorddata), the columns TimeStart and TimeEnd are returned with format DD/MM/YYYY, where DD are days, MM are months and YYYY are years.

Examples

if (FALSE) {
# Link to country list
L1 <- linkGenerator(type = "locations",
                    addDefault = "false",
                    includeDependencies = "false",
                    includeFormerCountries = "false")
L1

# Link to location types (for Egypt)
# With strings rather than codes
L2 <- linkGenerator(type = "locAreaTypes",
                    indicatorTypeIds = "Population by sex",
                    locIds = "Egypt",
                    isComplete = "Abridged")
L2

# Link to subgroup types (for Egypt)
L3 <- linkGenerator(type = "subGroups",
                    indicatorTypeIds = 8,
                    locIds = 818,
                    isComplete = 0)
L3

# Link to indicator list
L4 <- linkGenerator(type = "indicators",
                    addDefault = "false")
L4

# Link to data process type list
L5 <- linkGenerator(type = "dataProcessTypes")
L5
}